prevent navigation to profile and search page if user is not logged in
